Reorganized code inside server folder
All code was related to the server, no need to stay outside of it
